removing old manually-generated n2 diagrams; fixing circuit test stuff to be able generate an n2
