+++ openmdao/test_suite/test_examples/test_circuit_analysis.py
@@ -4,104 +4,7 @@ import numpy as np
 import unittest
 
 from openmdao.api import ExplicitComponent, ImplicitComponent, Group, NewtonSolver, DirectSolver
-
 from openmdao.utils.assert_utils import assert_rel_error
-
-
-class Resistor(ExplicitComponent):
-    """Computes current across a resistor using Ohm's law."""
-
-    def initialize(self):
-        self.options.declare('R', default=1., desc='Resistance in Ohms')
-
-    def setup(self):
-        self.add_input('V_in', units='V')
-        self.add_input('V_out', units='V')
-        self.add_output('I', units='A')
-
-        self.declare_partials('I', 'V_in', method='fd')
-        self.declare_partials('I', 'V_out', method='fd')
-
-    def compute(self, inputs, outputs):
-        deltaV = inputs['V_in'] - inputs['V_out']
-        outputs['I'] = deltaV / self.options['R']
-
-
-class Diode(ExplicitComponent):
-    """Computes current across a diode using the Shockley diode equation."""
-
-    def initialize(self):
-        self.options.declare('Is', default=1e-15, desc='Saturation current in Amps')
-        self.options.declare('Vt', default=.025875, desc='Thermal voltage in Volts')
-
-    def setup(self):
-        self.add_input('V_in', units='V')
-        self.add_input('V_out', units='V')
-        self.add_output('I', units='A')
-
-        self.declare_partials('I', 'V_in', method='fd')
-        self.declare_partials('I', 'V_out', method='fd')
-
-    def compute(self, inputs, outputs):
-        deltaV = inputs['V_in'] - inputs['V_out']
-        Is = self.options['Is']
-        Vt = self.options['Vt']
-        outputs['I'] = Is * np.exp(deltaV / Vt - 1)
-
-
-class Node(ImplicitComponent):
-    """Computes voltage residual across a node based on incoming and outgoing current."""
-
-    def initialize(self):
-        self.options.declare('n_in', default=1, types=int, desc='number of connections with + assumed in')
-        self.options.declare('n_out', default=1, types=int, desc='number of current connections + assumed out')
-
-    def setup(self):
-        self.add_output('V', val=5., units='V')
-
-        for i in range(self.options['n_in']):
-            i_name = 'I_in:{}'.format(i)
-            self.add_input(i_name, units='A')
-
-        for i in range(self.options['n_out']):
-            i_name = 'I_out:{}'.format(i)
-            self.add_input(i_name, units='A')
-
-        #note: we don't declare any partials wrt `V` here,
-        #      because the residual doesn't directly depend on it
-        self.declare_partials('V', 'I*', method='fd')
-
-    def apply_nonlinear(self, inputs, outputs, residuals):
-        residuals['V'] = 0.
-        for i_conn in range(self.options['n_in']):
-            residuals['V'] += inputs['I_in:{}'.format(i_conn)]
-        for i_conn in range(self.options['n_out']):
-            residuals['V'] -= inputs['I_out:{}'.format(i_conn)]
-
-
-# note: This is defined twice in the file. Once so you can import it, and once inside a test that gets included in the docs.
-class Circuit(Group):
-
-    def setup(self):
-        self.add_subsystem('n1', Node(n_in=1, n_out=2), promotes_inputs=[('I_in:0', 'I_in')])
-        self.add_subsystem('n2', Node())  # leaving defaults
-
-        self.add_subsystem('R1', Resistor(R=100.), promotes_inputs=[('V_out', 'Vg')])
-        self.add_subsystem('R2', Resistor(R=10000.))
-        self.add_subsystem('D1', Diode(), promotes_inputs=[('V_out', 'Vg')])
-
-        self.connect('n1.V', ['R1.V_in', 'R2.V_in'])
-        self.connect('R1.I', 'n1.I_out:0')
-        self.connect('R2.I', 'n1.I_out:1')
-
-        self.connect('n2.V', ['R2.V_out', 'D1.V_in'])
-        self.connect('R2.I', 'n2.I_in:0')
-        self.connect('D1.I', 'n2.I_out:0')
-
-        self.nonlinear_solver = NewtonSolver()
-        self.nonlinear_solver.options['iprint'] = 2
-        self.nonlinear_solver.options['maxiter'] = 20
-        self.linear_solver = DirectSolver()
 
 
 class TestCircuit(unittest.TestCase):
@@ -110,7 +13,7 @@ class TestCircuit(unittest.TestCase):
 
         from openmdao.api import Group, NewtonSolver, DirectSolver, Problem, IndepVarComp
 
-        from openmdao.test_suite.test_examples.test_circuit_analysis import Resistor, Diode, Node
+        from openmdao.test_suite.scripts.circuit_analysis import Resistor, Diode, Node
 
         class Circuit(Group):
 
@@ -170,7 +73,7 @@ class TestCircuit(unittest.TestCase):
 
         from openmdao.api import Group, NewtonSolver, DirectSolver, Problem, IndepVarComp
 
-        from openmdao.test_suite.test_examples.test_circuit_analysis import Resistor, Diode, Node
+        from openmdao.test_suite.scripts.circuit_analysis import Resistor, Diode, Node
 
         class Circuit(Group):
 
@@ -226,7 +129,7 @@ class TestCircuit(unittest.TestCase):
 
         from openmdao.api import Problem, IndepVarComp
 
-        from openmdao.test_suite.test_examples.test_circuit_analysis import Circuit
+        from openmdao.test_suite.scripts.circuit_analysis import Circuit
 
         p = Problem()
         model = p.model
@@ -259,7 +162,7 @@ class TestCircuit(unittest.TestCase):
     def test_circuit_advanced_newton(self):
         from openmdao.api import ArmijoGoldsteinLS, Problem, IndepVarComp
 
-        from openmdao.test_suite.test_examples.test_circuit_analysis import Circuit
+        from openmdao.test_suite.scripts.circuit_analysis import Circuit
 
         p = Problem()
         model = p.model
@@ -301,7 +204,7 @@ class TestCircuit(unittest.TestCase):
         from openmdao.api import ArmijoGoldsteinLS, Problem, IndepVarComp, BalanceComp, ExecComp
         from openmdao.api import NewtonSolver, DirectSolver, NonlinearRunOnce, LinearRunOnce
 
-        from openmdao.test_suite.test_examples.test_circuit_analysis import Circuit
+        from openmdao.test_suite.scripts.circuit_analysis import Circuit
 
         p = Problem()
         model = p.model
